# docs/openapi.yml
{
  "openapi": "3.0.2",
  "info": {
    "title": "Observer API",
    "description": "Observer API server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/people": {
      "post": {
        "tags": [
          "people"
        ],
        "summary": "Create Person",
        "operationId": "create_person_people_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/people/{person_id}": {
      "get": {
        "tags": [
          "people"
        ],
        "summary": "Get Person",
        "operationId": "get_person_people__person_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Person Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "person_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "people"
        ],
        "summary": "Update Person",
        "operationId": "update_person_people__person_id__put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Person Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "person_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "people"
        ],
        "summary": "Delete Person",
        "operationId": "delete_person_people__person_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Person Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "person_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Forwarded-For",
              "type": "string",
              "default": ""
            },
            "name": "X-Forwarded-For",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/people/{person_id}/personal-info": {
      "get": {
        "tags": [
          "people"
        ],
        "summary": "Get Personal Info",
        "operationId": "get_personal_info_people__person_id__personal_info_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Person Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "person_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/people/{person_id}/migration-records": {
      "get": {
        "tags": [
          "people",
          "migration history"
        ],
        "summary": "Get Person Migration Records",
        "description": "Get migration records for a person",
        "operationId": "get_person_migration_records_people__person_id__migration_records_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Person Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "person_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Person Migration Records People  Person Id  Migration Records Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FullMigrationHistoryResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NewPersonRequest": {
        "title": "NewPersonRequest",
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DisplacedPersonStatus"
              }
            ],
            "description": "Current status",
            "default": "registered"
          },
          "reference_id": {
            "title": "Reference Id",
            "type": "string",
            "description": "Reference ID, maybe some of state issued IDs etc."
          },
          "email": {
            "title": "Email",
            "type": "string",
            "description": "Contact email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string",
            "description": "Full name"
          },
          "sex": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Sex"
              }
            ],
            "description": "Person's sex"
          },
          "pronoun": {
            "title": "Pronoun",
            "type": "string",
            "description": "Person's pronouns"
          },
          "birth_date": {
            "title": "Birth Date",
            "type": "string",
            "description": "Birth date",
            "format": "date-time"
          },
          "notes": {
            "title": "Notes",
            "type": "string",
            "description": "Additional notes"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string",
            "description": "Primary phone number"
          },
          "phone_number_additional": {
            "title": "Phone Number Additional",
            "type": "string",
            "description": "Displaced person ID"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Related project ID"
          },
          "category_id": {
            "title": "Category Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Vulnerability category ID"
          },
          "office_id": {
            "title": "Office Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Registration office ID"
          },
          "tags": {
            "title": "Tags",
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of tags"
          }
        }
      },
...
      "Sex": {
        "title": "Sex",
        "enum": [
          "male",
          "female",
          "unknown"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
    }
  }
}